requirements.txt
requests~=2.32.3
tldextract~=5.1.2
PyYAML~=6.0.2